Change the way we write to the dashboard
Previously, we would have to add calls to update the dashboard in every
command. We would often forget to do this. It also didn't make sense for
some subsystems where there wasn't a continually running default
command.

This change removes the need to add calls to commands to update the
dashboard and instead will call the updateDashboard method for every
subsystem in the various periodic functions. This also has the benefit
of writing data to the dashboard while the robot is disabled.