Refactor intake
I have refactored the intake subsystem and its default command so that
there is only one method in Intake (wheelSpeed) controls the wheels and
the command doesn't have to make the decision about whether to
dynamically vary the speed based on sensor information.

I have also cleaned up the condition under which dynamic speed variation
is used (cube in "sweet spot"...not too close, not too far).

I have added a constant to allow us to turn dynamic mode off easily for
testing or if there is a problem with sensors.

I made the getDistance commands private to test the idea that we don't
need that information outside of the subsystem...we only need to know if
we have a cube or not.

A few things were renamed for clarity.